inference/requirement.txt
python-dotenv=1.0.1